backend/requirements.txt
# ============================================================================
# Python Dependencies for E-commerce Realtime Pipeline
# ============================================================================

# Event Generator Dependencies
kafka-python==2.0.2        # Kafka producer client

# Spark Dependencies
pyspark==3.5.0             # Spark structured streaming

# Database Dependencies (for testing/debugging)
psycopg2-binary==2.9.9     # PostgreSQL adapter
